+++ src/senaite/storage/interfaces.py
@@ -22,7 +22,6 @@ from bika.lims.interfaces import IBikaLIMS
 from senaite.core.interfaces import ISenaiteCatalogObject
 from senaite.lims.interfaces import ISenaiteLIMS
 from zope.interface import Interface
-from zope.viewlet.interfaces import IViewletManager
 
 
 class ISenaiteStorageLayer(IBikaLIMS, ISenaiteLIMS):
@@ -77,11 +76,6 @@ class IStorageContainer(IStorageContent):
 class IStorageSamplesContainer(IStorageContent):
     """Marker interface for objects that represent a type of storage container
     designed for the storage of multiple samples (storage box, tube rack, etc.)
-    """
-
-
-class IStorageJS(IViewletManager):
-    """A viewlet manager that provides the JavaScripts for DataBox
     """