+++ CHANGES.rst
@@ -4,6 +4,7 @@ Changelog
 2.7.0 (unreleased)
 ------------------
 
+- #63 Result type-specific controls in retention rules settings
 - #64 Add 'Storage Expiry Date' column in samples listing, under 'Stored'
 - #62 Add configurable storage retention period
 - #61 Added StorageManager and StorageAssistant roles and counterpart groups